requirements.txt
Flask==2.1.0
gunicorn==20.1.0
google-cloud-datastore==2.8.1
opentelemetry-api==1.12.0
opentelemetry-sdk==1.12.0
opentelemetry-exporter-gcp-trace==1.3.0
